project-chronos/requirements.txt
numpy>=1.22.0
pydantic>=2.0.0
pyyaml>=6.0
scipy>=1.9.0
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
httpx>=0.24.0
